test(provider): move LSP call graph provider tests to separate module
The comprehensive unit tests for the LSP call graph provider were relocated from `src/tests/mod.rs` to a new file `src/tests/provider.rs`. This refactoring improves test code organization and maintainability by isolating provider-specific tests.
